Python-Backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.2.1
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chromadb==0.6.3
chroma-hnswlib==0.7.6
click==8.1.7
coloredlogs==15.0.1
Deprecated==1.2.15
dnspython==2.7.0
durationpy==0.9
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.6
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.5
kubernetes==31.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
numba==0.58.1
oauthlib==3.2.2
onnxruntime==1.20.1
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
overrides==7.7.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
posthog==3.7.4
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic>=2.9.0,<3.0.0
pydantic_core==2.14.6
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyJWT==2.10.1
python-multipart==0.0.19
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.11.3
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
zipp==3.21.0
pypdf[full]==5.1.0
python-docx==0.8.11
beautifulsoup4==4.12.2
markdown==3.5.1
python-pptx==0.6.21
openpyxl==3.1.2
lxml==5.3.0
pandas==2.2.3
pytz==2024.2
pillow==11.0.0
soupsieve==2.6
openai==1.58.1
distro==1.9.0
nest_asyncio==1.5.6
hypercorn==0.14.3
toml==0.10.2
h2==4.1.0
hyperframe==6.0.1
hpack==4.0.0
http3==0.6.7
h11==0.14.0
httpcore==1.0.7
sentence-transformers==3.3.1
threadpoolctl==3.5.0
joblib==1.4.2
scipy==1.15.1
httpx==0.28.0
priority==2.0.0
wsproto==1.2.0
jiter==0.8.2
langchain==0.3.13
langchain-text-splitters==0.3.4
langchain_core==0.3.28
langsmith==0.2.3
requests_toolbelt==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain_community==0.3.13
tiktoken==0.8.0
regex==2024.11.6
langchain-openai==0.2.14
langchain-chroma==0.2.1
psutil==6.1.1
ollama==0.4.4
docx2txt==0.8
yt-dlp==2024.12.23
webvtt-py==0.4.6
langchain-ollama==0.2.2
openai-whisper==20240930
accelerate>=0.20.3
bitsandbytes>=0.41.1
safetensors>=0.4.0
llvmlite==0.43.0
einops==0.8.0
optimum==1.23.3
datasets==3.2.0
pyarrow==18.1.0
multiprocess==0.70.17
dill>=0.3.6
aiohttp==3.11.11
multidict==6.1.0
attrs>=23.1.0
yarl==1.18.3
propcache==0.2.1
async-timeout==5.0.1
aiohappyeyeballs==2.4.4
aiosignal==1.3.2
frozenlist==1.5.0
xxhash==3.5.0
diskcache==5.6.3
hqq==0.2.2
termcolor==2.5.0
langchain-huggingface==0.1.2